Added error checking

+++ django_project/galleries/views.py
@@ -28,12 +28,11 @@ class GalleryRetrieve(generics.RetrieveAPIView):
 
 @api_view(["POST"])
 def create_or_update_gallery(request):
-    print('request', request)
-    print('request data', request.data)
     data = request.data.copy()
-    print('data', data)
     id_exists = data.get("id", None)
     if id_exists is None:
+        if "name" not in data or "description" not in data or "layout" not in data:
+            return JsonResponse({"response": "Name, description, or layout does not exist."}, status=400)
         gallery_serializer_class = GallerySerializer(data={"name": data["name"], "description": data["description"], "layout": data["layout"]})
         if gallery_serializer_class.is_valid():
             gallery = gallery_serializer_class.save()
@@ -42,8 +41,15 @@ def create_or_update_gallery(request):
         id = gallery.id
     else:
         id = id_exists
-        Gallery.objects.get(id = id).images.all().delete()
+        referenced_gallery = Gallery.objects.filter(id=id)
+        if len(referenced_gallery) == 0:
+            return JsonResponse({"response": "Gallery does not exist."}, status=400)
+        Gallery.objects.get(id=id).images.all().delete()
+    if "images" not in data:
+        return JsonResponse({"response": "Images does not exist."}, status=400) 
     for index, image in enumerate(data["images"]):
+        if "caption" not in image or "url" not in image:
+            return JsonResponse({"response": "Caption or image url does not exist."}, status=400)
         image.update({"gallery": id, "index": index})
         image["img_url"] = image.pop("url")
         image_serializer_class = ImageSerializer(data=image)
